+++ src/components/SidePanel/SidePanel.tsx
@@ -9,8 +9,8 @@ import { SidePanel } from 'pc-nrfconnect-shared';
 
 import AdvancedOptions from './AdvancedOptions';
 import Instructions from './Instructions';
+import PowerProfilerParams from './PowerProfilerParams';
 import TraceCollector from './Tracing/TraceCollector';
-import TraceConverter from './Tracing/TraceConverter';
 import TraceFileInformation from './Tracing/TraceFileInformation';
 
 import './sidepanel.scss';
@@ -20,8 +20,8 @@ export default () => (
     <SidePanel className="side-panel">
         <TraceCollector />
         <Instructions />
-        <TraceConverter />
         <TraceFileInformation />
+        <PowerProfilerParams />
         <AdvancedOptions />
     </SidePanel>
 );