+++ src/features/tracing/__mocks__/@nordicsemiconductor/nrf-monitor-lib-js.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
+export default { start: jest.fn().mockReturnValue(1), stop: jest.fn() };
+
 const getPluginsDir = jest.fn().mockReturnValue('mocked_plugin_dir_path');
 
-export default { start: jest.fn().mockReturnValue(1), stop: jest.fn() };
 export { getPluginsDir };